# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TodoAttributes": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"task": {
						"type": "string"
					},
					"isDone": {
						"type": "boolean"
					},
					"dueDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"userId",
					"task",
					"dueDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_TodoAttributes.userId-or-task-or-dueDate_": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"task": {
						"type": "string"
					},
					"dueDate": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					}
				},
				"required": [
					"userId",
					"task",
					"dueDate"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"TodoAttributesCreationParams": {
				"$ref": "#/components/schemas/Pick_TodoAttributes.userId-or-task-or-dueDate_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "typescript-express-template",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/todos/{todoId}": {
			"get": {
				"operationId": "GetTodo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TodoAttributes"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "todoId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "task",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/todos": {
			"post": {
				"operationId": "CreateTodo",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TodoAttributesCreationParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}